refactor(org-scope): move config and table repositories
