+++ api/src/repositories/__init__.py
@@ -1,6 +1,7 @@
 # Data access layer - PostgreSQL repositories
 from src.repositories.base import BaseRepository
 from src.repositories.cli_sessions import CLISessionRepository
+from src.repositories.config import ConfigRepository
 from src.repositories.data_providers import DataProviderRepository
 from src.repositories.execution_logs import ExecutionLogRepository
 from src.repositories.executions import (
@@ -14,6 +15,7 @@ from src.core.exceptions import AccessDeniedError
 from src.repositories.oauth import OAuthProviderRepository, OAuthTokenRepository
 from src.repositories.org_scoped import OrgScopedRepository
 from src.repositories.organizations import OrganizationRepository
+from src.repositories.tables import TableRepository
 from src.repositories.users import UserRepository
 from src.repositories.workflows import WorkflowRepository
 
@@ -21,6 +23,7 @@ __all__ = [
     "AccessDeniedError",
     "BaseRepository",
     "CLISessionRepository",
+    "ConfigRepository",
     "DataProviderRepository",
     "ExecutionLogRepository",
     "ExecutionRepository",
@@ -32,6 +35,7 @@ __all__ = [
     "OAuthTokenRepository",
     "OrgScopedRepository",
     "OrganizationRepository",
+    "TableRepository",
     "UserRepository",
     "WorkflowRepository",
     # Standalone functions for workers/consumers
