+++ api/src/routers/config.py
@@ -7,27 +7,20 @@ Uses OrgScopedRepository for standardized org scoping.
 """
 
 import logging
-from datetime import datetime, timezone
-from typing import Any
 from uuid import UUID
 
 from fastapi import APIRouter, HTTPException, Query, status
-from sqlalchemy import select
 
 # Import existing Pydantic models for API compatibility
 from src.models import (
     ConfigResponse,
-    ConfigType,
     SetConfigRequest,
     UpdateConfigRequest,
 )
 
 from src.core.auth import Context, CurrentSuperuser
-from src.core.log_safety import log_safe
-from src.core.org_filter import resolve_org_filter, OrgFilterType
-from src.models import Config as ConfigModel
-from src.models.enums import ConfigType as ConfigTypeEnum
-from src.repositories.org_scoped import OrgScopedRepository
+from src.core.org_filter import resolve_org_filter
+from src.repositories.config import ConfigRepository
 
 # Import cache functions
 try:
@@ -41,355 +34,6 @@ except ImportError:
 logger = logging.getLogger(__name__)
 
 router = APIRouter(tags=["Configuration"])
-
-
-# =============================================================================
-# Repository (using OrgScopedRepository)
-# =============================================================================
-
-
-class ConfigRepository(OrgScopedRepository[ConfigModel]):  # type: ignore[type-var]
-    """
-    Config repository using OrgScopedRepository.
-
-    Configs use the CASCADE scoping pattern:
-    - Org-specific configs + global (NULL org_id) configs
-
-    Config is SDK-only (no RBAC), so role_table is not set.
-    """
-
-    model = ConfigModel
-    role_table = None  # No RBAC for configs
-
-    async def list_configs(
-        self,
-        filter_type: OrgFilterType = OrgFilterType.ORG_PLUS_GLOBAL,
-    ) -> list[ConfigResponse]:
-        """List configs with specified filter type."""
-        from src.models.orm.integrations import Integration
-        from sqlalchemy import and_
-
-        query = select(self.model, Integration.name.label("integration_name")).outerjoin(
-            Integration,
-            and_(
-                self.model.integration_id == Integration.id,
-                Integration.is_deleted.is_(False),
-            )
-        )
-
-        # Apply filter based on filter_type
-        if filter_type == OrgFilterType.ALL:
-            pass
-        elif filter_type == OrgFilterType.GLOBAL_ONLY:
-            query = query.where(self.model.organization_id.is_(None))
-        elif filter_type == OrgFilterType.ORG_ONLY:
-            query = query.where(self.model.organization_id == self.org_id)
-        else:
-            # Inline cascade scope to avoid type mismatch with joined query
-            from sqlalchemy import or_
-            if self.org_id is not None:
-                query = query.where(
-                    or_(
-                        self.model.organization_id == self.org_id,
-                        self.model.organization_id.is_(None),
-                    )
-                )
-            else:
-                query = query.where(self.model.organization_id.is_(None))
-
-        query = query.order_by(self.model.key)
-
-        result = await self.session.execute(query)
-        rows = result.all()
-
-        schemas = []
-        for row in rows:
-            c = row[0]  # ConfigModel
-            integration_name = row[1]  # Integration.name or None
-
-            raw_value = c.value.get("value") if isinstance(c.value, dict) else c.value
-            # Mask secret values in list responses
-            if c.config_type == ConfigTypeEnum.SECRET:
-                display_value = "[SECRET]"
-            else:
-                display_value = raw_value
-
-            schemas.append(
-                ConfigResponse(
-                    id=c.id,
-                    key=c.key,
-                    value=display_value,
-                    type=ConfigType(c.config_type.value) if c.config_type else ConfigType.STRING,
-                    scope="org" if c.organization_id else "GLOBAL",
-                    org_id=str(c.organization_id) if c.organization_id else None,
-                    integration_id=str(c.integration_id) if c.integration_id else None,
-                    integration_name=integration_name,
-                    description=c.description,
-                    updated_at=c.updated_at,
-                    updated_by=c.updated_by,
-                )
-            )
-        return schemas
-
-    async def get_config(self, key: str) -> ConfigModel | None:
-        """Get config by key with cascade scoping: org-specific > global.
-
-        Uses base class get() which prioritizes org-specific over global
-        to avoid MultipleResultsFound when the same key exists in both scopes.
-        """
-        return await self.get(key=key)
-
-    async def merged_for_sdk(self) -> dict[str, Any]:
-        """Return the full merged config dict for this scope.
-
-        Replaces ``ConfigResolver.load_config_for_scope``. The merged shape
-        is ``{key: {"value": v, "type": t}, ...}`` where org-specific entries
-        override globals on key collision.
-
-        Cache is read-through with the versioned-key invalidation introduced
-        in Phase 5 (see ``api/src/core/cache/keys.py::config_hash_key_versioned``).
-
-        Excludes integration-scoped configs (``integration_id IS NOT NULL``) —
-        those are accessed via the integration system, not ``config.get()``.
-        """
-        import json
-        from typing import Any
-
-        from src.core.cache.keys import (
-            TTL_CONFIG,
-            config_hash_key_versioned,
-        )
-        from src.core.cache.redis_client import get_shared_redis
-        from src.core.log_safety import log_safe
-
-        org_id_str = str(self.org_id) if self.org_id is not None else None
-
-        # Read-through cache.
-        try:
-            r = await get_shared_redis()
-            hash_key = await config_hash_key_versioned(r, org_id_str)
-            cached = await r.hgetall(hash_key)  # type: ignore[misc]
-            if cached:
-                out: dict[str, Any] = {}
-                for k, v in cached.items():
-                    ks = k.decode() if isinstance(k, bytes) else k
-                    vs = v.decode() if isinstance(v, bytes) else v
-                    try:
-                        out[ks] = json.loads(vs)
-                    except json.JSONDecodeError:
-                        out[ks] = {"value": vs, "type": "string"}
-                return out
-        except Exception as e:  # cache is best-effort
-            logger.warning(f"Config cache read failed: {e}")
-
-        # Cache miss — load from DB via cascade.
-        # Globals first, then org-specific overrides (org wins on key collision).
-        config_dict: dict[str, Any] = {}
-
-        global_q = select(self.model).where(
-            self.model.organization_id.is_(None),
-            self.model.integration_id.is_(None),
-        )
-        global_rows = (await self.session.execute(global_q)).scalars()
-        for cfg in global_rows:
-            v = cfg.value.get("value") if isinstance(cfg.value, dict) else cfg.value
-            config_dict[cfg.key] = {
-                "value": v,
-                "type": cfg.config_type.value if cfg.config_type else "string",
-            }
-
-        if self.org_id is not None:
-            org_q = select(self.model).where(
-                self.model.organization_id == self.org_id,
-                self.model.integration_id.is_(None),
-            )
-            org_rows = (await self.session.execute(org_q)).scalars()
-            for cfg in org_rows:
-                v = cfg.value.get("value") if isinstance(cfg.value, dict) else cfg.value
-                config_dict[cfg.key] = {
-                    "value": v,
-                    "type": cfg.config_type.value if cfg.config_type else "string",
-                }
-
-        # Write-through cache (best-effort).
-        if config_dict:
-            try:
-                r = await get_shared_redis()
-                hash_key = await config_hash_key_versioned(r, org_id_str)
-                mapping = {k: json.dumps(v) for k, v in config_dict.items()}
-                await r.hset(hash_key, mapping=mapping)  # type: ignore[misc]
-                await r.expire(hash_key, TTL_CONFIG)
-            except Exception as e:
-                logger.warning(f"Config cache write failed: {e}")
-
-        logger.debug(
-            f"Loaded {len(config_dict)} config entries for org={log_safe(org_id_str)}"
-        )
-        return config_dict
-
-    async def get_config_strict(self, key: str) -> ConfigModel | None:
-        """Get config strictly in current org scope (no fallback)."""
-        query = select(self.model).where(
-            self.model.key == key,
-            self.model.organization_id == self.org_id,
-        )
-        result = await self.session.execute(query)
-        return result.scalar_one_or_none()
-
-    async def set_config(self, request: SetConfigRequest, updated_by: str) -> ConfigResponse:
-        """Create or update a config in current org scope."""
-        now = datetime.now(timezone.utc)
-
-        # Handle secret encryption if this is a SECRET type
-        stored_value = request.value
-        if request.type == ConfigType.SECRET:
-            from src.core.security import encrypt_secret
-            stored_value = encrypt_secret(request.value)
-
-        # Build value object for JSONB storage
-        config_value = {
-            "value": stored_value,
-        }
-
-        # Convert API ConfigType to DB ConfigTypeEnum
-        # Both enums have same values, so we can use the value to lookup
-        db_config_type = ConfigTypeEnum(request.type.value) if request.type else ConfigTypeEnum.STRING
-
-        # Check if config exists in current org scope
-        existing = await self.get_config_strict(request.key)
-
-        if existing:
-            # Update existing
-            existing.value = config_value
-            existing.config_type = db_config_type
-            existing.description = request.description
-            existing.updated_at = now
-            existing.updated_by = updated_by
-            await self.session.flush()
-            await self.session.refresh(existing)
-            config = existing
-        else:
-            # Create new
-            config = ConfigModel(
-                key=request.key,
-                value=config_value,
-                config_type=db_config_type,
-                description=request.description,
-                organization_id=self.org_id,
-                created_at=now,
-                updated_at=now,
-                updated_by=updated_by,
-            )
-            self.session.add(config)
-            await self.session.flush()
-            await self.session.refresh(config)
-
-        logger.info(f"Set config {log_safe(request.key)} in org {self.org_id}")
-
-        # Extract value from JSONB for response
-        stored_value = config.value.get("value") if isinstance(config.value, dict) else config.value
-        return ConfigResponse(
-            id=config.id,
-            key=config.key,
-            value=stored_value,
-            type=request.type if request.type else ConfigType.STRING,
-            scope="org" if config.organization_id else "GLOBAL",
-            org_id=str(config.organization_id) if config.organization_id else None,
-            description=config.description,
-            updated_at=config.updated_at,
-            updated_by=config.updated_by,
-        )
-
-    async def update_config_by_id(
-        self,
-        config_id: UUID,
-        request: UpdateConfigRequest,
-        updated_by: str,
-    ) -> tuple[ConfigResponse, UUID | None, str] | None:
-        """Update a config by ID. Allows changing organization scope.
-
-        For SECRET type configs, if value is None or empty string, the existing
-        encrypted value is preserved.
-
-        Returns ``(response, old_org_id, old_key)`` so the caller can
-        invalidate the previous cache entry when ``organization_id`` or
-        ``key`` changed in the same update. Returns ``None`` if the row
-        does not exist.
-        """
-        query = select(self.model).where(self.model.id == config_id)
-        result = await self.session.execute(query)
-        config = result.scalar_one_or_none()
-        if not config:
-            return None
-
-        # Snapshot pre-update identity so the router can invalidate the
-        # old cache entry if either field changes (rename or org-move).
-        old_org_id = config.organization_id
-        old_key = config.key
-
-        now = datetime.now(timezone.utc)
-
-        # Determine the effective type (use request type if provided, else keep existing)
-        effective_type = request.type if request.type is not None else (
-            ConfigType(config.config_type.value) if config.config_type else ConfigType.STRING
-        )
-
-        # Handle value update — for secrets, empty/None means keep existing
-        if request.value is not None and request.value != "":
-            stored_value = request.value
-            if effective_type == ConfigType.SECRET:
-                from src.core.security import encrypt_secret
-                stored_value = encrypt_secret(request.value)
-            config.value = {"value": stored_value}
-        elif effective_type != ConfigType.SECRET:
-            # Non-secret types: empty string is a valid value, None means no change
-            if request.value is not None:
-                config.value = {"value": request.value}
-
-        if request.key is not None:
-            config.key = request.key
-        if request.type is not None:
-            config.config_type = ConfigTypeEnum(request.type.value)
-        if "description" in (request.model_fields_set or set()):
-            config.description = request.description
-        if "organization_id" in (request.model_fields_set or set()):
-            config.organization_id = request.organization_id
-        config.updated_at = now
-        config.updated_by = updated_by
-        await self.session.flush()
-        await self.session.refresh(config)
-
-        logger.info(f"Updated config {log_safe(config.key)} (id={log_safe(config_id)}) org={log_safe(config.organization_id)}")
-
-        response_type = ConfigType(config.config_type.value) if config.config_type else ConfigType.STRING
-        stored_value = config.value.get("value") if isinstance(config.value, dict) else config.value
-        response = ConfigResponse(
-            id=config.id,
-            key=config.key,
-            value=stored_value,
-            type=response_type,
-            scope="org" if config.organization_id else "GLOBAL",
-            org_id=str(config.organization_id) if config.organization_id else None,
-            description=config.description,
-            updated_at=config.updated_at,
-            updated_by=config.updated_by,
-        )
-        return response, old_org_id, old_key
-
-    async def delete_config(self, config_id: UUID) -> ConfigModel | None:
-        """Delete config by ID. Returns the deleted config or None if not found."""
-        query = select(self.model).where(self.model.id == config_id)
-        result = await self.session.execute(query)
-        config = result.scalar_one_or_none()
-        if not config:
-            return None
-
-        key = config.key
-        await self.session.delete(config)
-        await self.session.flush()
-
-        logger.info(f"Deleted config {log_safe(key)} (id={log_safe(config_id)})")
-        return config
 
 
 # =============================================================================
