+++ api/tests/unit/test_org_scoping_enforcement.py
@@ -77,10 +77,6 @@ ALLOW_LIST_INLINE_ORG: set[tuple[str, str, str]] = {
     ('routers/cli.py', 'ConfigModel.organization_id == org_uuid,', 'cli config inline; phase 5 migrates'),
     ('routers/cli.py', 'Table.organization_id == org_uuid,', 'cli_create_table exact-scope uniqueness check (NOT cascade)'),
     # cli_list_tables migrated to TableRepository.list() in phase 6.
-    ('routers/config.py', 'query = query.where(self.model.organization_id.is_(None))', 'ConfigRepository cascade override; phase 5 absorbs into base'),
-    ('routers/config.py', 'query = query.where(self.model.organization_id == self.org_id)', 'ConfigRepository cascade override; phase 5 absorbs into base'),
-    ('routers/config.py', 'self.model.organization_id == self.org_id,', 'ConfigRepository cascade override; phase 5 absorbs into base'),
-    ('routers/config.py', 'self.model.organization_id.is_(None),', 'ConfigRepository cascade override; phase 5 absorbs into base'),
     ('routers/executions.py', 'query = query.where(ExecutionModel.organization_id == org_id)', 'Execution identity-entity filter (permanent)'),
     ('routers/export_import.py', 'Config.organization_id == mapping.organization_id', 'manifest sync inline; phase 8 follow-up'),
     ('routers/export_import.py', 'else Config.organization_id.is_(None),', 'manifest sync inline; phase 8 follow-up'),
@@ -121,9 +117,6 @@ ALLOW_LIST_INLINE_ORG: set[tuple[str, str, str]] = {
     ('routers/roi_reports.py', 'query = query.where(WorkflowROIDaily.organization_id == org_uuid)', 'identity-entity scope filter (permanent)'),
     ('routers/roi_reports.py', '.join(Organization, ExecutionMetricsDaily.organization_id == Organization.id)', 'identity-entity scope filter (permanent)'),
     ('routers/roles.py', 'KnowledgeNamespaceRoleORM.organization_id == entry.organization_id,', 'KnowledgeNamespaceRole identity-entity filter (permanent)'),
-    ('routers/tables.py', 'query = query.where(self.model.organization_id.is_(None))', 'TableRepository cascade override; phase 6 relocates'),
-    ('routers/tables.py', 'query = query.where(self.model.organization_id == self.org_id)', 'TableRepository cascade override; phase 6 relocates'),
-    ('routers/tables.py', 'self.model.organization_id == self.org_id,', 'TableRepository cascade override; phase 6 relocates'),
     ('routers/tables.py', 'CustomClaimORM.organization_id == organization_id', 'tables custom claim cross-ref; phase 6 migrates'),
     ('routers/tools.py', 'query = query.where(Workflow.organization_id == filter_org_id)', 'tools workflow list; phase 6 migrates'),
     ('routers/tools.py', 'query = query.where(Workflow.organization_id.is_(None))', 'tools workflow list; phase 6 migrates'),
